Attempt to fix agents CPU lock and log parsing (#13)
We have observed several issues that may be related:

1. Agents behave erractly when dealing with certain log payloads. A
common observation here is that steps that produce a large volume of
logs will cause some steps to be stuck "pending" forever.

2. Agents use way more CPU than should be expected, we often see 200-300
millicores of CPU per Workflow per agent.

3. We commonly see Agents displaying thousands of error lines such as
those, often with the same timestamp, indicating an error with parsing
logs and attempting to perform GRPC requests, which may explain issues 1
and 2.

```
{"level":"error","error":"rpc error: code = Internal desc = grpc: error while marshaling: string field contains invalid UTF-8","time":"2024-04-05T21:32:25Z","caller":"/src/agent/rpc/client_grpc.go:335","message":"grpc error: log(): code: Internal"}
{"level":"error","error":"rpc error: code = Internal desc = grpc: error while marshaling: string field contains invalid UTF-8","time":"2024-04-05T21:32:25Z","caller":"/src/agent/rpc/client_grpc.go:335","message":"grpc error: log(): code: Internal"}
{"level":"error","error":"rpc error: code = Internal desc = grpc: error while marshaling: string field contains invalid UTF-8","time":"2024-04-05T21:32:25Z","caller":"/src/agent/rpc/client_grpc.go:335","message":"grpc error: log(): code: Internal"}
```

4. We've also observed that agents will sometimes "drop out" of the
worker queue, as we often seem to have less active workers than we
should as observed by both the `woodpecker_worker_count` and the actual
maximum number of tasks.

Seeing that the issue points to `client_grpc.go:335`, this pull request
attempts to fix these problems by:

1. Removing codes.Internal from being a retryable GRPC status. Now agent
GRPC calls that fail with codes.Internal (such as those above) will not
be retried. I've searched and there's not an agreement on what GRPC
codes should be retried but Internal does not seem to be a common one to
retry anyway.

2. Add a timeout of 30 seconds to any retries. Currently, the
exponential retries have a maximum timeout of _15 minutes_. I think this
is behind the huge cpu increase as some pods might be locked forever
retrying these requests. Of course, since we have completely removed the
possibility of a retry for these particular error codes, this should not
make any difference for this scenario... but hopefully will solve the
issue for other similar potential scenarios.